Refactor vocab view out of content view
