+++ Travel KZH/Travel KZH/Views/ContentView.swift
@@ -21,34 +21,7 @@ struct ContentView: View {
             }
             // MARK: - NavigationView - Vocabulary
             NavigationView {
-                Form {
-                    Section {
-                        let glossary: [Glossary] = [.greetings, .conversations, .dining, .food, .weather]
-                        
-                        List(glossary, children: \.terms) { term in
-                            
-                            if let englishTranslation = term.englishTranslation {
-                                HStack {
-                                    Text(englishTranslation)
-                                        .frame(minWidth: 0, maxWidth: .infinity, alignment: .leading)
-                                        .font(Font.system(size: 16, design: .default))
-                                    
-                                    Spacer()
-                                    Text(term.russianTranslation!)
-                                        .frame(minWidth: 0, maxWidth: .infinity, alignment: .leading)
-                                        .font(Font.system(size: 16, design: .monospaced))
-                                }
-                            } else {
-                                Image(systemName: term.icon)
-                                Text(term.categoryName)
-                            }
-                            
-                        }
-                    } header: {
-                        Text("Basic words and sentences")
-                    }
-                }
-                .navigationTitle("Kazakhstan 🇰🇿")
+                VocabularyView()
             }
             .tabItem {
                 Image(systemName: "person.2.wave.2.fill")
